hands-on-labs/01-snowflake-migration-lab/01-setup-snowflake/producer/requirements.txt
snowflake-connector-python==3.7.0
cryptography>=41.0.0   # for key-pair auth support
